+++ magenpy/__init__.py
@@ -16,7 +16,7 @@ from magenpy.simulation.GWASimulator import GWASimulator
 
 from magenpy.utils.data_utils import *
 
-__version__ = '0.0.7'
+__version__ = '0.0.8'
 
 
 config = configparser.ConfigParser()