+++ magenpy/stats/variant/utils.py
@@ -36,7 +36,7 @@ def compute_allele_frequency_plink2(genotype_matrix, temp_dir='temp'):
 
     freq_df = pd.read_csv(plink_output + ".afreq", delim_whitespace=True)
     freq_df.rename(columns={'ID': 'SNP', 'ALT': 'A1', 'ALT_FREQS': 'MAF'}, inplace=True)
-    merged_df = merge_snp_tables(genotype_matrix.get_snp_table(['SNP', 'A1']), freq_df)
+    merged_df = merge_snp_tables(genotype_matrix.get_snp_table(['SNP', 'A1', 'A2']), freq_df)
 
     if len(merged_df) != genotype_matrix.n_snps:
         raise ValueError("Length of allele frequency table does not match number of SNPs.")